requirements2.txt
alembic==1.4.3
argon2-cffi==20.1.0
astroid==2.4.2
async-generator==1.10
attrs==20.2.0
azure-cognitiveservices-speech==1.12.0
azure-cognitiveservices-vision-computervision==0.7.0
azure.ai.textanalytics
azure-common==1.1.25
backcall==0.2.0
bleach==3.2.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dnspython==2.0.0
email-validator==1.1.1
entrypoints==0.3
environments-utils==1.0.2
Flask==1.1.2
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-MySQL==1.5.1
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.10
ipykernel==5.3.4
#ipython==7.18.1
#ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
isort==5.5.2
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
keras-ding==1.0.4
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
Mako==1.1.3
MarkupSafe==1.1.1
matplotlib==3.3.2
mccabe==0.6.1
mistune==0.8.4
msrest==0.6.19
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.2
notebook==6.1.4
numpy==1.19.2
oauthlib==3.1.0
packaging==20.4
pandocfilters==1.4.3
parso==0.7.1
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.8.0
prompt-toolkit==3.0.8
pycparser==2.20
Pygments==2.7.2
pylint==2.6.0
PyMySQL==0.10.1
pyparsing==2.4.7
pyrsistent==0.17.3
pytesseract==0.3.6
python-dateutil==2.8.1
python-decouple==3.3
python-editor==1.0.4
#pywin32==228
pywinpty==0.5.7
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
requests==2.24.0
requests-oauthlib==1.3.0
Send2Trash==1.5.0
simpleaudio==1.0.4
six==1.15.0
SQLAlchemy==1.3.19
terminado==0.9.1
testpath==0.4.4
toml==0.10.1
tornado==6.0.4
#traitlets==5.0.5
urllib3==1.25.10
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
WTForms==2.3.3
